Updated element deletion logic
