+++ src/main/java/ua/knd12/controller/StudentController.java
@@ -1,41 +1,4 @@
 package ua.knd12.controller;
 
-import ua.knd12.model.User;
-
-import java.util.Arrays;
-
 public class StudentController {
-    protected static User[]initialArray;
-    private int counter = 0;
-
-    public StudentController(User[] initialArray) {
-        this.initialArray = initialArray;
-    }
-
-
-    public void add(User u) {
-
-        int newSize = (initialArray.length*2)+1;
-        if (initialArray.length<=counter) {
-            initialArray = Arrays.copyOf(initialArray, newSize);
-        }
-        initialArray[counter] = u;
-        counter++;
-    }
-
-    public void delete(int id) {
-        int pos = initialArray.length-1;
-        for (int i=0;i<initialArray.length;i++) {
-            if(initialArray[i].getId()==id) {
-                initialArray[pos] = initialArray[i];
-                initialArray[i] = null;
-                counter--;
-                pos--;
-            }
-        }
-    }
-
-    public static User[] getInitialArray() {
-        return initialArray;
-    }
 }
