+++ src/main/java/ua/knd12/model/User.java
@@ -4,11 +4,12 @@ public abstract class User {
     private int id;
     private String name;
     private String surname;
+    private static int counter = 0;
 
     public User(String name, String surname, int id) {
         this.name = name;
         this.surname = surname;
-        this.id = id++;
+        this.id = counter++;
     }
 
     public int getId() {
