+++ src/main/java/ua/knd12/viewer/ApplicationDisplay.java
@@ -1,13 +1,11 @@
 package ua.knd12.viewer;
 
-import ua.knd12.controller.StudentController;
 import ua.knd12.model.User;
-
-import java.util.Arrays;
+import ua.knd12.service.impl.UserServiceImpl;
 
 public class ApplicationDisplay {
     public void getUsers() {
-        for (User user : StudentController.getInitialArray()) {
+        for (User user : UserServiceImpl.getInitialArray()) {
             if (user!=null)
                 System.out.println(user);
         }
